refacto of environment names wip
