+++ src/super_gradients/common/crash_handler/exception_monitoring_setup.py
@@ -3,7 +3,7 @@ import logging
 import atexit
 
 from super_gradients.common.auto_logging.console_logging import ConsoleSink
-from super_gradients.common.environment.env_helpers import multi_process_safe, is_distributed
+from super_gradients.common.environment.ddp_utils import multi_process_safe, is_distributed
 from super_gradients.common.crash_handler.exception import ExceptionInfo
 
 try:
