+++ src/super_gradients/common/environment/__init__.py
@@ -2,6 +2,6 @@
 This module is in charge of environment variables and consts.
 """
 from super_gradients.common.environment.environment_config import device_config
-from super_gradients.common.environment.env_helpers import init_trainer, is_distributed
+from super_gradients.common.environment.ddp_utils import init_trainer, is_distributed
 
 __all__ = ["device_config", "init_trainer", "is_distributed"]
