+++ src/super_gradients/common/environment/utils.py
@@ -1,0 +1,75 @@
+import argparse
+import importlib
+import os
+import sys
+from typing import Any
+
+from omegaconf import OmegaConf
+
+from super_gradients.common.environment import environment_config
+
+
+def get_cls(cls_path):
+    """
+    A resolver for Hydra/OmegaConf to allow getting a class instead on an instance.
+    usage:
+    class_of_optimizer: ${class:torch.optim.Adam}
+    """
+    module = ".".join(cls_path.split(".")[:-1])
+    name = cls_path.split(".")[-1]
+    importlib.import_module(module)
+    return getattr(sys.modules[module], name)
+
+
+def get_environ_as_type(environment_variable_name: str, default=None, cast_to_type: type = str) -> object:
+    """
+    Tries to get an environment variable and cast it into a requested type.
+    :return: cast_to_type object, or None if failed.
+    :raises ValueError: If the value could not be casted into type 'cast_to_type'
+    """
+    value = os.environ.get(environment_variable_name, default)
+    if value is not None:
+        try:
+            return cast_to_type(value)
+        except Exception as e:
+            print(e)
+            raise ValueError(
+                f"Failed to cast environment variable {environment_variable_name} to type {cast_to_type}: the value {value} is not a valid {cast_to_type}"
+            )
+    return
+
+
+def hydra_output_dir_resolver(ckpt_root_dir, experiment_name):
+    if ckpt_root_dir is None:
+        output_dir_path = environment_config.PKG_CHECKPOINTS_DIR + os.path.sep + experiment_name
+    else:
+        output_dir_path = ckpt_root_dir + os.path.sep + experiment_name
+    return output_dir_path
+
+
+def register_hydra_resolvers():
+    """Register all the hydra resolvers required for the super-gradients recipes."""
+    OmegaConf.register_new_resolver("hydra_output_dir", hydra_output_dir_resolver, replace=True)
+    OmegaConf.register_new_resolver("class", lambda *args: get_cls(*args), replace=True)
+    OmegaConf.register_new_resolver("add", lambda *args: sum(args), replace=True)
+    OmegaConf.register_new_resolver("cond", lambda boolean, x, y: x if boolean else y, replace=True)
+    OmegaConf.register_new_resolver("getitem", lambda container, key: container[key], replace=True)  # get item from a container (list, dict...)
+    OmegaConf.register_new_resolver("first", lambda lst: lst[0], replace=True)  # get the first item from a list
+    OmegaConf.register_new_resolver("last", lambda lst: lst[-1], replace=True)  # get the last item from a list
+
+
+EXTRA_ARGS = []
+
+
+def pop_arg(arg_name: str, default_value: Any = None) -> Any:
+    """Get the specified args and remove them from argv"""
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument(f"--{arg_name}", default=default_value)
+    args, _ = parser.parse_known_args()
+
+    # Remove the ddp args to not have a conflict with the use of hydra
+    for val in filter(lambda x: x.startswith(f"--{arg_name}"), sys.argv):
+        EXTRA_ARGS.append(val)
+        sys.argv.remove(val)
+    return vars(args)[arg_name]
