+++ src/super_gradients/training/utils/checkpoint_utils.py
@@ -4,6 +4,7 @@ import pkg_resources
 
 import torch
 
+import super_gradients.common.environment.ddp_utils
 from super_gradients.common.abstractions.abstract_logger import get_logger
 from super_gradients.common import explicit_params_validation, ADNNModelRepositoryDataInterfaces
 from super_gradients.training.pretrained_models import MODEL_URLS
@@ -29,7 +30,7 @@ def get_checkpoints_dir_path(experiment_name: str, ckpt_root_dir: str = None):
     if ckpt_root_dir:
         return os.path.join(ckpt_root_dir, experiment_name)
     elif os.path.exists(environment_config.PKG_CHECKPOINTS_DIR):
-        return os.path.join(environment_config.PKG_CHECKPOINTS_DIR, experiment_name)
+        return os.path.join(super_gradients.common.environment.ddp_utils.PKG_CHECKPOINTS_DIR, experiment_name)
     else:
         raise ValueError("Illegal checkpoints directory: pass ckpt_root_dir that exists, or add 'checkpoints' to resources.")
